FastAPI/requirements.txt
# --- FASTAPI STACK ---

# Modern, fast (high-performance) web framework for building APIs with Python 3.7+ based on standard Python type hints.
fastapi==0.103.1

# Lightweight ASGI server to run FastAPI apps in development and production.
uvicorn==0.23.2

# --- DATABASE & ORM ---

# SQL toolkit and ORM that gives full power and flexibility of SQL in Python.
sqlalchemy==2.0.21

# Pure Python MySQL client, needed for SQLAlchemy to connect to MySQL or MariaDB.
pymysql==1.1.0

# --- SECURITY & AUTHENTICATION ---

# Cryptography library used for encryption, decryption, key generation, and digital signatures.
cryptography==41.0.4

# JWT (JSON Web Token) implementation in Python with cryptography support, used for token-based authentication.
python-jose[cryptography]==3.3.0

# Password hashing and verification library, includes bcrypt scheme.
passlib[bcrypt]==1.7.4

# Optional: explicit bcrypt dependency, required by passlib internally.
bcrypt>=4.0.1

# --- PRODUCTION DEPLOYMENT ---

# Production-grade WSGI HTTP server for Python web applications.
gunicorn==21.2.0

# --- CODE QUALITY & TESTING ---

# Static analysis tool to enforce coding standards and catch errors early.
pylint==3.3.7

# Powerful testing framework for writing unit, functional, and integration tests.
pytest==8.3.5

# Asynchronous HTTP client used for integration testing or making external API calls.
httpx==0.27.0

# --- CONFIGURATION MANAGEMENT ---

# Loads environment variables from a .env file into the application environment.
python-dotenv==1.0.1

# --- DATA VALIDATION ---

# Data validation and settings management using Python type hints; includes email validators.
pydantic[email]==2.5.3

# --- FORM HANDLING ---

# Required for handling form data in FastAPI, including file uploads and form submissions
python-multipart==0.0.20